pyproject.toml
[project]
name = "pyvbox"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "setuptools>=69.5.1",
    "pip>=24.0",
    "click>=8.1.7",
    "pytest>=8.1.1",
    "black>=24.4.0",
    "flake8>=7.0.0",
    "coverage>=7.4.4",
    "psutil>=5.9.8",
    "texttable>=1.7.0",
    "py-cpuinfo>=9.0.0",
    "lowkit @ git+https://github.com/terminal-labs-bem/lowkit.git",
    "highkit @ git+https://github.com/terminal-labs-bem/highkit.git",
    "dllist>=1.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"pyvbox" = "pyvbox:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pyvbox"]
